+++ contracts/stake/CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Improved performance of get_provisioners [#1447]
+- Adjusted assert for transfer caller [#1235]
+- Deployed 3 transfer contracts in test [#1235]
 - Change dependencies declarations enforce bytecheck [#1371]
 - Removed 'phoenix-core' dependency [#1138]
 
@@ -22,5 +24,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [#1447]: https://github.com/dusk-network/rusk/issues/1448
 [#1447]: https://github.com/dusk-network/rusk/issues/1447
+[#1235]: https://github.com/dusk-network/rusk/issues/1235
 [#1371]: https://github.com/dusk-network/rusk/issues/1371
 [#1138]: https://github.com/dusk-network/rusk/issues/1138